Move LockEditor
